Mejoro para que solo yo pueda modificar mierdas

+++ controllers/JuegosController.php
@@ -51,46 +51,58 @@ class JuegosController extends Controller
 
     public function actionCreate()
     {
-        $model = new Juegos();
-        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
-            Yii::$app->response->format = Response::FORMAT_JSON;
-            return ActiveForm::validate($model);
-        }
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+        if (Yii::$app->user->identity->nombre == 'josesabor') {
+            $model = new Juegos();
+            if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
+                Yii::$app->response->format = Response::FORMAT_JSON;
+                return ActiveForm::validate($model);
+            }
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['index']);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+                'totalG' => Generos::lista(),
+                'totalC' => Consolas::lista(),
+            ]);
+        } else {
             return $this->redirect(['index']);
         }
-
-        return $this->render('create', [
-            'model' => $model,
-            'totalG' => Generos::lista(),
-            'totalC' => Consolas::lista(),
-        ]);
     }
 
     public function actionUpdate($id)
     {
-        $model = $this->findJuego($id);
-        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
-            Yii::$app->response->format = Response::FORMAT_JSON;
-            return ActiveForm::validate($model);
-        }
+        if (Yii::$app->user->identity->nombre == 'josesabor') {
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            $model = $this->findJuego($id);
+            if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post())) {
+                Yii::$app->response->format = Response::FORMAT_JSON;
+                return ActiveForm::validate($model);
+            }
+
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['index']);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+                'totalG' => Generos::lista(),
+                'totalC' => Consolas::lista(),
+            ]);
+        } else {
             return $this->redirect(['index']);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-            'totalG' => Generos::lista(),
-            'totalC' => Consolas::lista(),
-        ]);
     }
 
     public function actionDelete($id)
     {
-        $model = $this->findJuego($id);
-        $model->delete();
-        Yii::$app->session->setFlash('success', 'Fila borrada con éxito.');
+        Yii::debug(Yii::$app->user->identity->nombre);
+        if (Yii::$app->user->identity->nombre == 'josesabor') {
+            $model = $this->findJuego($id);
+            $model->delete();
+            Yii::$app->session->setFlash('success', 'Fila borrada con éxito.');
+        }
         return $this->redirect(['index']);
     }
